% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regions.R
\name{add_regions}
\alias{add_regions}
\title{Map countries to World Bank regions}
\usage{
add_regions(democracy_data = load_data())
}
\arguments{
\item{democracy_data}{Dataframe of countries.}
}
\value{
Input dataframe with added region column.
}
\description{
Map countries to World Bank regions
}
