% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regime_map.R
\name{make_regime_map}
\alias{make_regime_map}
\title{Gemerate an interactive leaflet map of countries colored by democratic regime, marked with regime history.}
\usage{
make_regime_map(democracy_data, world_data, include_markers = TRUE)
}
\arguments{
\item{democracy_data}{Dataset of countries' regimes to map}

\item{world_data}{Dataset of global country geometries from rnaturalearth package}

\item{include_markers}{default TRUE; includes clickable map markers with regime history}
}
\value{
A leaflet map
}
\description{
Gemerate an interactive leaflet map of countries colored by democratic regime, marked with regime history.
}
\examples{
dat <- load_data()
world_dat <- load_world_data()
make_regime_map(dat, world_dat)
}
