% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_coups.R
\name{plot_country_coups}
\alias{plot_country_coups}
\title{Create a plotly plot graphing the coup attempts of a certain country.}
\usage{
plot_country_coups(
  coup_data = load_coup_data(),
  democracy_data = load_data(),
  country_of_interest = "Bolivia",
  coup_code = verified
)
}
\arguments{
\item{coup_data}{Optional dataframe of coup d'etat data from}

\item{democracy_data}{Optional dataframe of countries' regimes.}

\item{country_of_interest}{character value with the full name of the country to plot}

\item{coup_code}{Optional column name of the coup coding scheme to use. Default is Powell and Thyne. See https://www.uky.edu/~clthyn2/coup_data/Codebook.pdf for more options.}
}
\value{
A plotly plot
}
\description{
Create a plotly plot graphing the coup attempts of a certain country.
}
\examples{
dat <- load_data()
coup_dat <- load_coup_data()
plot_country_coups(coup_dat, dat, country_of_interest="Haiti", coup_code=verified)
}
