% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_country_regimes.R
\name{summarize_country_regimes}
\alias{summarize_country_regimes}
\title{Summarize country regimes}
\usage{
summarize_country_regimes(democracy_data, country = "Bolivia")
}
\arguments{
\item{democracy_data}{Dataset of countries' regimes to summarize from.}

\item{country}{Country whose regimes should be summarized.}
}
\value{
A tibble with columns for year and new regime of each regime change.
}
\description{
Summarize country regimes
}
\examples{
dat <- load_data()
summarize_country_regimes(dat, "United States")
}
