% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_country}
\alias{validate_country}
\title{Validate a country}
\usage{
validate_country(country, democracy_data = load_data())
}
\arguments{
\item{country}{Three letter ISO country code.}

\item{democracy_data}{Optional dataframe of countries' regimes.}
}
\value{
Invisibly returns \code{TRUE} when the country code input is present in the data.
}
\description{
Validate a country
}
