% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_region}
\alias{validate_region}
\title{Validate a region}
\usage{
validate_region(region)
}
\arguments{
\item{region}{World Bank region.}
}
\value{
Invisibly returns \code{TRUE} when the region input is a valid World Bank region.
}
\description{
Validate a region
}
