% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_country_regimes.R
\name{summarize_country_regimes}
\alias{summarize_country_regimes}
\title{Display summary of a country's regimes}
\usage{
summarize_country_regimes(country, democracy_data = load_data())
}
\arguments{
\item{country}{Three letter ISO country code.}

\item{democracy_data}{Optional dataframe of countries' regimes.}
}
\value{
A tibble with a row for each regime change the country has had
and columns for the year and new regime of the change.
}
\description{
Display summary of a country's regimes
}
\examples{
summarize_country_regimes("BOL")
}
