% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_region_regimes.R
\name{summarize_region_regimes}
\alias{summarize_region_regimes}
\title{Display summary of a region's regimes}
\usage{
summarize_region_regimes(region, democracy_data = load_data())
}
\arguments{
\item{region}{World Bank region. Options include: "East Asia & Pacific",
"Europe & Central Asia",
"Latin America & Caribbean",
"Middle East & North Africa",
"North America",
"South Asia",
"Sub-Saharan Africa"}

\item{democracy_data}{Optional dataframe of countries' regimes.}
}
\value{
A wide tibble with the proportions of regimes the region has add
that were of each type of governance.
}
\description{
Display summary of a region's regimes
}
\examples{
summarize_country_regimes("East Asia & Pacific")
}
