% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate a dataset}
\usage{
validate_dataset(country, democracy_data = load_data())
}
\arguments{
\item{democracy_data}{Dataframe of countries' regimes.}
}
\value{
Invisibly returns \code{TRUE} when the data input is a dataframe.
}
\description{
Validate a dataset
}
