% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dictatr_parquet.R
\name{write_dictatr_parquet}
\alias{write_dictatr_parquet}
\title{Write dictatR datasets as Parquet files}
\usage{
write_dictatr_parquet(path = ".")
}
\arguments{
\item{path}{A directory where the Parquet files should be saved.}
}
\value{
A character vector of the desired file path to write Parquet files.

A character vector of file paths to the written Parquet files.
}
\description{
Write dictatR datasets as Parquet files
}
